// server/migrations/snapshots/9d1746249de938e6ce7b744e7854084fd60cb028ee49b1ca9d957279ce84141d/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Artist: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly musicBrainzId: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Genre: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Rating: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly songId: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly value: CodecTypes['pg/int4@1']['output'];
    };
    readonly Song: {
      readonly acousticness: CodecTypes['pg/float8@1']['output'] | null;
      readonly artistId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly danceability: CodecTypes['pg/float8@1']['output'] | null;
      readonly duration: CodecTypes['pg/float8@1']['output'] | null;
      readonly energy: CodecTypes['pg/float8@1']['output'] | null;
      readonly externalId: CodecTypes['pg/text@1']['output'] | null;
      readonly genreId: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly instrumentalness: CodecTypes['pg/float8@1']['output'] | null;
      readonly listenCount: CodecTypes['pg/int8@1']['output'] | null;
      readonly listenerCount: CodecTypes['pg/int4@1']['output'] | null;
      readonly liveness: CodecTypes['pg/float8@1']['output'] | null;
      readonly musicBrainzId: CodecTypes['pg/text@1']['output'] | null;
      readonly popularity: CodecTypes['pg/float8@1']['output'] | null;
      readonly releaseYear: CodecTypes['pg/int4@1']['output'] | null;
      readonly speechiness: CodecTypes['pg/float8@1']['output'] | null;
      readonly tempo: CodecTypes['pg/float8@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly valence: CodecTypes['pg/float8@1']['output'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
    readonly UserFavoriteArtist: {
      readonly artistId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly UserFavoriteGenre: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly genreId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Artist: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly musicBrainzId: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Genre: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Rating: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly songId: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly value: CodecTypes['pg/int4@1']['input'];
    };
    readonly Song: {
      readonly acousticness: CodecTypes['pg/float8@1']['input'] | null;
      readonly artistId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly danceability: CodecTypes['pg/float8@1']['input'] | null;
      readonly duration: CodecTypes['pg/float8@1']['input'] | null;
      readonly energy: CodecTypes['pg/float8@1']['input'] | null;
      readonly externalId: CodecTypes['pg/text@1']['input'] | null;
      readonly genreId: CodecTypes['pg/int4@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly instrumentalness: CodecTypes['pg/float8@1']['input'] | null;
      readonly listenCount: CodecTypes['pg/int8@1']['input'] | null;
      readonly listenerCount: CodecTypes['pg/int4@1']['input'] | null;
      readonly liveness: CodecTypes['pg/float8@1']['input'] | null;
      readonly musicBrainzId: CodecTypes['pg/text@1']['input'] | null;
      readonly popularity: CodecTypes['pg/float8@1']['input'] | null;
      readonly releaseYear: CodecTypes['pg/int4@1']['input'] | null;
      readonly speechiness: CodecTypes['pg/float8@1']['input'] | null;
      readonly tempo: CodecTypes['pg/float8@1']['input'] | null;
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly valence: CodecTypes['pg/float8@1']['input'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
    };
    readonly UserFavoriteArtist: {
      readonly artistId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly UserFavoriteGenre: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly genreId: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Artist: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly musicBrainzId: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Genre: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Rating: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly songId: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly value: CodecTypes['pg/int4@1']['output'];
    };
    readonly Song: {
      readonly acousticness: CodecTypes['pg/float8@1']['output'] | null;
      readonly artistId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly danceability: CodecTypes['pg/float8@1']['output'] | null;
      readonly duration: CodecTypes['pg/float8@1']['output'] | null;
      readonly energy: CodecTypes['pg/float8@1']['output'] | null;
      readonly externalId: CodecTypes['pg/text@1']['output'] | null;
      readonly genreId: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly instrumentalness: CodecTypes['pg/float8@1']['output'] | null;
      readonly listenCount: CodecTypes['pg/int8@1']['output'] | null;
      readonly listenerCount: CodecTypes['pg/int4@1']['output'] | null;
      readonly liveness: CodecTypes['pg/float8@1']['output'] | null;
      readonly musicBrainzId: CodecTypes['pg/text@1']['output'] | null;
      readonly popularity: CodecTypes['pg/float8@1']['output'] | null;
      readonly releaseYear: CodecTypes['pg/int4@1']['output'] | null;
      readonly speechiness: CodecTypes['pg/float8@1']['output'] | null;
      readonly tempo: CodecTypes['pg/float8@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly valence: CodecTypes['pg/float8@1']['output'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
    readonly UserFavoriteArtist: {
      readonly artistId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly UserFavoriteGenre: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly genreId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;